Make SAML trust anchors work on verification of the SAML request
SAML uses a concept called trust anchors where you can say, I trust this party
because of $reasons. These trust anchors can be the issuing party, the DN of
the certificate, etc. The previous code implemented a partial solution to this,
it just said that the trust anchor was the DN of the certificate of the signer.
This is somewhat weird, as the certificate is already verified by the CA
certificate which is user supplied.

Now you can submit a CA and you can inject trust anchors into the SOAP binding
so it is checked with the verify_xml call. The trust anchors can be one of the
following `subject`, `issuer` or `issuer_hash`.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>