Verify the SAMLResponse based on the raw query string
BREAKING CHANGE: The verify function now requires the *raw* query string from
the server in order to verify the signature.

Because `URI` parses and re-encodes URI-escapes in uppercase (`%3f` becomes
`%3F`, for instance), which leads to signature verification failures if the
other party uses lower case (or mixed case). This seems to have been an issue
with the current code base as well due to the various flags that need to be
supplied to the constructor to work around it.

This code is mostly ported from Zaaksysteem's code base, written by MSTREEK.

Users of lighttpd need to be aware that they need to configure their instance
with the following http-parseopts:

    server.http-parseopts = (
        "url-normalize"            => "disable",
        "url-normalize-unreserved" => "disable",
        "url-normalize-required"   => "disable"
    )

You cannot change it on a URI basis because lighttpd needs a parsed URI before
it can process conditions as was mentioned on #lighttpd:

> The request must be parsed (using server.http-parseopts settings) *before* you
> can apply lighttpd.conf conditions to the parsed results.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>